chore: improve loading state on events table
